+++ changelog.md
@@ -5,6 +5,7 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add support for JSON5 and JSONC configuration files.
 - YAML and XML configuration support is now optional. You need to install the `click_extra[yaml]` and `click_extra[xml]` extra dependency groups to enable it.
 - Move all custom types to `click_extra._types` module.
 - Avoid importing all types at runtime to reduce startup time.